add icons again for future use. break screenController.js down into more modules
