+++ src/index.js
@@ -1,6 +1,6 @@
 import "./assets/styles/normalise.css";
 import "./assets/styles/global.css";
-import { createScreenController } from "./controllers/Screen";
+import createScreenController from "./controllers/Screen";
 
 // localStorage.clear(); // comment out when you need to reset
 createScreenController();
